==> workshops/2024_EFA_SRU/03_get_data_in_shape.ipynb <==
import pandas as pd
import os
import wget
import PyPDF2
from PyPDF2 import PdfReader
import urllib.parse

### Test: Einlesen einer Seite eines PDFs

def test_read_pdf(file_path, file_name):
    reader = PdfReader(file_path + "/" + file_name)
    number_of_pages = len(reader.pages)
    page = reader.pages[10]
    text = page.extract_text()
    print(text)

test_read_pdf("downloads", "1192182693.pdf")

### Einlesen einer kompletten PDF-Datei:

def read_pdf_content(file_path, filename):
    try:
        reader = PdfReader(file_path + "/" + filename)
        number_of_pages = len(reader.pages)
        text = ""
        for page in reader.pages:
            text += page.extract_text()  # + "\n"
        return text
    except Exception as e:
        print(f"Fehler beim Einlesen der Datei {filename}: {e}")
        return "Fehler beim Einlesen der Datei"

test = read_pdf_content("downloads", "1192182693.pdf")
print(test)

#### Einlesen aller Dateien aus dem "downloads"-Verzeichnis, die auf .pdf enden. Anschließend werden die Texte im Zusammenhang mit ihrem Dateinamen in ein Pandas_Dataframe überführt:

def get_text_from_pdfs(file_path):
    pdf_text = {}
    for filename in os.listdir(file_path):
        if filename.endswith(".pdf"):
            text = read_pdf_content(file_path, filename)
            pdf_text[filename] = text
    return pdf_text

pdf_texts = get_text_from_pdfs("downloads")

print(pdf_texts)

#Löschen der ganzen Zeilenumbrüche ("\n"): 
pdf_texts = {key: value.replace('\n', ' ') for key, value in pdf_texts.items()}


for key, value in pdf_texts.items():
    print(f"Key: {key}, Type of value: {type(value)}")
    if value is None:
        print(f"None value found for key: {key}")

#Umwandeln in ein Dataframe:
df_texts = pd.DataFrame.from_dict(pdf_texts, orient='index', columns=['Text'])
df_texts.reset_index(inplace=True)
df_texts.columns = ['filename', 'text']
df_texts

df_texts.to_hdf("pdf_texts.h5", key="df_texts")
